backend/requirements.txt
streamlit
pandas
networkx
matplotlib
openpyxl
plotly

# Database 
SQLAlchemy
psycopg2-binary

# Testing
pytest
pytest-asyncio
pytest-cov
factory-boy
faker
fastapi
uvicorn
pandas
networkx
pytest
python-multipart
python-jose
passlib[bcrypt]
